Squashed commit of the following:
commit d2ddc8f0f5309b10aa11b2ddcb77e7bdada7f395
Merge: c9e7f47 42e5a4e
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Mon Jul 28 16:05:51 2025 +0300

    Merge pull request #12 from actingweb/changelog_fixes

    Correct changelog

commit 42e5a4e1ff4d08ad88191b9a9859da103440e67b
Author: Greger Wedel <greger@greger.io>
Date:   Mon Jul 28 15:05:25 2025 +0200

    Correct changelog

commit c9e7f470a17bf189dcee413a6e52b2809912d943
Merge: 53b8286 c7f8daf
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Mon Jul 28 12:34:15 2025 +0300

    Merge pull request #11 from actingweb/bug_fixes

    Fix trustee_root and forms data bugs

commit c7f8daf869e7178cf8811faeac972b412185a167
Author: Greger Wedel <greger@greger.io>
Date:   Mon Jul 28 11:33:25 2025 +0200

    Fix trustee_root and forms data bugs

commit 53b8286e913ef57ea64ae4a941af57221e6eff3f
Merge: 8f57464 a8ef1ae
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Mon Jul 28 10:43:43 2025 +0300

    Merge pull request #10 from actingweb/fastapi_support

    Fastapi_support

commit a8ef1ae8166a846c6467591a608fa67d155c8c76
Author: Greger Wedel <greger@greger.io>
Date:   Mon Jul 28 09:34:20 2025 +0200

    fastapi support

commit 0a55450def5a48298a369e25b3657dccced4d8e8
Author: Greger Wedel <greger@greger.io>
Date:   Mon Jul 28 09:31:53 2025 +0200

    Clean up coverage files

commit 8f57464c9e809185cfeebdb7d6485e32c55cb816
Merge: 3594d57 9c8f16f
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Sat Jul 26 22:45:12 2025 +0300

    Merge pull request #9 from actingweb/remove_legacy_interface

    Refactored out legacy dev interface

commit 9c8f16f03cfd3bfc7311171f66923a273fe76798
Author: Greger Wedel <greger@greger.io>
Date:   Sat Jul 26 21:40:15 2025 +0200

    Refactored out legacy dev interface

commit 3594d57114448b16be9697840c65532956af0fd7
Merge: 818a5cd 32131af
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Sat Jul 26 14:56:03 2025 +0300

    Merge pull request #8 from actingweb/minor_fixes

    Minor_fixes

commit 32131afcd4d7128f22a7d7f78eec7c5eb400e478
Author: Greger Wedel <greger@greger.io>
Date:   Sat Jul 26 13:53:02 2025 +0200

    Minor fixes

commit 7dd38c7e5b2f1f5317076adfdbd28391cc8dab9d
Author: Greger Wedel <greger@greger.io>
Date:   Sat Jul 26 11:02:32 2025 +0200

    Clean up gitignore

commit 818a5cd0959d936bedb76a17f3274043ca98ab52
Merge: e1c8f66 4c21983
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Thu Jul 17 13:18:03 2025 +0200

    Merge pull request #6 from actingweb/claude_refactor

    Claude_refactor

commit 4c219830ee8ba67a08749b5165c57239fe22f47f
Author: Greger Wedel <greger@greger.io>
Date:   Thu Jul 17 13:16:41 2025 +0200

    Reevamp changelog and bump to 3.0.1

commit cba1a6f0c1c24f294132c253bb8af46861aeaf33
Author: Greger Wedel <greger@greger.io>
Date:   Wed Jul 16 10:25:17 2025 +0200

    Add callback support for /methods and /actions

commit 275c9238f05a5cb92d27b1316c97b1b57e942ec7
Author: Greger Wedel <greger@greger.io>
Date:   Wed Jul 16 09:43:39 2025 +0200

    Refactored developer interface

commit d5c85e7afbc26b6a2559c23d1b658f03c4ac730e
Author: Greger Wedel <greger@greger.io>
Date:   Tue Jul 15 16:32:52 2025 +0200

    Fix pypi

commit 8a3244f86a3d5b8c6418165d4ce22cae6a5d01e2
Author: Greger Wedel <greger@greger.io>
Date:   Tue Jul 15 16:02:41 2025 +0200

    Remove all legacy code for non-standard urlfetch lib in appengine

commit 09ea7538ece8c2c137f6b85bd660ac61ea63a0ac
Author: Greger Wedel <greger@greger.io>
Date:   Tue Jul 15 12:53:05 2025 +0200

    Fixed missing None checks

commit f4db0e4980f3ed9a20d304a189a7e10d5fb36c88
Author: Greger Wedel <greger@greger.io>
Date:   Mon Jul 14 14:40:53 2025 +0200

    Remaining type annotations

commit 5feabafea6b9695b7e7d8f154bc311e8d969a9e3
Author: Greger Wedel <greger@greger.io>
Date:   Mon Jul 14 13:57:39 2025 +0200

    Refactor to remove Google AppEngine legacy code and the migration support from 2.5.0

commit c582421877842aa80389cc7ff7d4fa1b0488056d
Author: Greger Wedel <greger@greger.io>
Date:   Sun Jul 13 10:49:56 2025 +0200

    Update gitignore

commit e79b1104351c2d27052d26a8560b53c5bf13ba93
Author: Greger Wedel <greger@greger.io>
Date:   Sun Jul 13 10:41:48 2025 +0200

    Modernizing using claude

commit 2214c603ee434fd40a151ca993fae0c2ca61e41e
Author: Greger Wedel <greger@greger.io>
Date:   Fri Jul 11 13:29:00 2025 +0200

    Claude modernization

commit 10e6f5bbe111fcb937cca4cffec36254a7b0d6b1
Author: Greger Wedel <greger@greger.io>
Date:   Fri Jul 11 10:53:34 2025 +0200

    Claude init

commit e1c8f66451f547c920c64c4e2a702698e3a0d299
Merge: 5517db4 98bb515
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Thu Apr 22 20:22:52 2021 +0200

    Merge pull request #5 from actingweb:fix-sub-diff-dynamodb

    Fix to subscription diffs

commit 98bb5158d45a4767b7c68cb5547bc6a9b15b1b35
Author: Greger Wedel <greger@greger.io>
Date:   Thu Apr 22 20:22:23 2021 +0200

    Fix to subscription diffs

commit 5517db4cecff674f4625b544ce6d2cb3483f7341
Merge: 34655c2 b16395d
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Sun Apr 11 11:11:45 2021 +0200

    Merge pull request #4 from actingweb:fix_2_6_3_version_error_in_release

    Fix 2.6.3 versioning problem

commit b16395d4e896f634552edd9540e69c3beb12a0ca
Author: Greger Wedel <greger@greger.io>
Date:   Sun Apr 11 11:11:13 2021 +0200

    Fix 2.6.3 versioning problem

commit 34655c2c9ab82d281b3e28d17635b5649603c7bf
Merge: a02ded7 e48784b
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Sun Apr 11 11:02:21 2021 +0200

    Merge pull request #3 from actingweb:fix-oauth-security-refresh-issue

    Fix peertrustee error with count in dynamodb

commit e48784b9661f548b99a33c33d76fa902e4036e57
Author: Greger Wedel <greger@greger.io>
Date:   Sun Apr 11 10:59:01 2021 +0200

    Fix peertrustee error with count in dynamodb

commit a02ded7d48912fda9646f8df9adc030eeae3334a
Merge: 313c4c5 25292f0
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Tue Oct 20 20:23:44 2020 +0200

    Merge pull request #2 from actingweb/fix-oauth-security-refresh-issue

    Fix oauth security issue on refresh

commit 25292f08a2022424dddb0da73cbce4681063e6e2
Author: Greger Wedel <greger@greger.io>
Date:   Tue Oct 20 20:15:56 2020 +0200

    Fix

commit 313c4c5cfe3f85eb04748a0a90e596d4f34a4039
Author: Greger Wedel <greger@greger.io>
Date:   Sun Aug 30 10:29:59 2020 +0200

    Fix token refresh to use basic auth header

commit 380548599df1cfe17c59ec3211f1bd43fd101ce1
Merge: 042775a b470b1b
Author: Greger Teigre Wedel <greger@greger.io>
Date:   Wed Aug 26 21:01:13 2020 +0200

    Merge pull request #1 from actingweb/fitbit_oauth2

    Support basic auth header with clientId and secret

commit b470b1b20105af2a3ae8e874e85d353aa9e7fc95
Author: Greger Wedel <greger@greger.io>
Date:   Wed Aug 26 21:00:09 2020 +0200

    Support basic auth header with clientId and secret

commit 042775a3efa500c305327348d5ef1ff5f76cfce5
Author: Greger Wedel <greger.wedel@cxense.com>
Date:   Sat Jun 22 09:04:21 2019 +0200

    Fix new repo link

+++ actingweb/subscription.py
@@ -1,45 +1,59 @@
-from builtins import object
 import datetime
 import logging
+from typing import Any
 
 
-class Subscription(object):
+class Subscription:
     """Base class with core subscription methods (storage-related)"""
 
-    def get(self):
-        """Retrieve subscription from db given pre-initialized variables """
+    def get(self) -> dict[str, Any]:
+        """Retrieve subscription from db given pre-initialized variables"""
         if not self.actor_id or not self.peerid or not self.subid:
-            return None
+            return {}
         if self.subscription and len(self.subscription) > 0:
             return self.subscription
-        self.subscription = self.handle.get(actor_id=self.actor_id,
-                                            peerid=self.peerid,
-                                            subid=self.subid)
+        if self.handle:
+            self.subscription = self.handle.get(
+                actor_id=self.actor_id, peerid=self.peerid, subid=self.subid
+            )
+        else:
+            self.subscription = {}
         if not self.subscription:
             self.subscription = {}
         return self.subscription
 
-    def create(self, target=None, subtarget=None, resource=None, granularity=None, seqnr=1):
+    def create(
+        self, target: str | None = None, subtarget: str | None = None, resource: str | None = None, granularity: str | None = None, seqnr: int = 1
+    ) -> bool:
         """Create new subscription and push it to db"""
         if self.subscription and len(self.subscription) > 0:
-            logging.debug("Attempted creation of subscription when already loaded from storage")
+            logging.debug(
+                "Attempted creation of subscription when already loaded from storage"
+            )
             return False
         if not self.actor_id or not self.peerid:
-            logging.debug("Attempted creation of subscription without actor_id or peerid set")
+            logging.debug(
+                "Attempted creation of subscription without actor_id or peerid set"
+            )
             return False
         if not self.subid:
             now = datetime.datetime.utcnow()
-            seed = self.config.root + now.strftime("%Y%m%dT%H%M%S%f")
-            self.subid = self.config.new_uuid(seed)
-        if not self.handle.create(actor_id=self.actor_id,
-                                  peerid=self.peerid,
-                                  subid=self.subid,
-                                  granularity=granularity,
-                                  target=target,
-                                  subtarget=subtarget,
-                                  resource=resource,
-                                  seqnr=seqnr,
-                                  callback=self.callback):
+            if self.config:
+                seed = self.config.root + now.strftime("%Y%m%dT%H%M%S%f")
+                self.subid = self.config.new_uuid(seed)
+            else:
+                self.subid = None
+        if not self.handle or not self.handle.create(
+            actor_id=self.actor_id,
+            peerid=self.peerid,
+            subid=self.subid,
+            granularity=granularity,
+            target=target,
+            subtarget=subtarget,
+            resource=resource,
+            seqnr=seqnr,
+            callback=self.callback,
+        ):
             return False
         self.subscription["id"] = self.actor_id
         self.subscription["subscriptionid"] = self.subid
@@ -63,7 +77,9 @@ class Subscription(object):
 
     def increase_seq(self):
         if not self.handle:
-            logging.debug("Attempted increase_seq without subscription retrieved from storage")
+            logging.debug(
+                "Attempted increase_seq without subscription retrieved from storage"
+            )
             return False
         self.subscription["sequence"] += 1
         return self.handle.modify(seqnr=self.subscription["sequence"])
@@ -73,15 +89,22 @@ class Subscription(object):
         if not self.actor_id or not self.subid or not blob:
             logging.debug("Attempted add_diff without actorid, subid, or blob")
             return False
+        if not self.config:
+            return False
         diff = self.config.DbSubscriptionDiff.DbSubscriptionDiff()
-        diff.create(actor_id=self.actor_id,
-                    subid=self.subid,
-                    diff=blob,
-                    seqnr=self.subscription["sequence"]
-                    )
+        diff.create(
+            actor_id=self.actor_id,
+            subid=self.subid,
+            diff=blob,
+            seqnr=self.subscription["sequence"],
+        )
         if not self.increase_seq():
-            logging.error("Failed increasing sequence number for subscription " +
-                          self.subid + " for peer " + self.peerid)
+            logging.error(
+                "Failed increasing sequence number for subscription "
+                + self.subid
+                + " for peer "
+                + self.peerid
+            )
         return diff.get()
 
     def get_diff(self, seqnr=0):
@@ -90,29 +113,42 @@ class Subscription(object):
             return None
         if not isinstance(seqnr, int):
             return None
+        if not self.config:
+            return None
         diff = self.config.DbSubscriptionDiff.DbSubscriptionDiff()
         return diff.get(actor_id=self.actor_id, subid=self.subid, seqnr=seqnr)
 
     def get_diffs(self):
         """Get all the diffs available for this subscription ordered by the timestamp, oldest first"""
+        if not self.config:
+            return []
         diff_list = self.config.DbSubscriptionDiff.DbSubscriptionDiffList()
         return diff_list.fetch(actor_id=self.actor_id, subid=self.subid)
 
     def clear_diff(self, seqnr):
         """Clears one specific diff"""
+        if not self.config:
+            return False
         diff = self.config.DbSubscriptionDiff.DbSubscriptionDiff()
         diff.get(actor_id=self.actor_id, subid=self.subid, seqnr=seqnr)
         return diff.delete()
 
     def clear_diffs(self, seqnr=0):
         """Clear all diffs up to and including a seqnr"""
+        if not self.config:
+            return False
         diff_list = self.config.DbSubscriptionDiff.DbSubscriptionDiffList()
         diff_list.fetch(actor_id=self.actor_id, subid=self.subid)
         diff_list.delete(seqnr=seqnr)
 
-    def __init__(self, actor_id=None, peerid=None, subid=None, callback=False, config=None):
+    def __init__(
+        self, actor_id=None, peerid=None, subid=None, callback=False, config=None
+    ):
         self.config = config
-        self.handle = self.config.DbSubscription.DbSubscription()
+        if self.config:
+            self.handle = self.config.DbSubscription.DbSubscription()
+        else:
+            self.handle = None
         self.subscription = {}
         if not actor_id:
             return
@@ -124,19 +160,19 @@ class Subscription(object):
             self.get()
 
 
-class Subscriptions(object):
-    """ Handles all subscriptions of a specific actor_id
+class Subscriptions:
+    """Handles all subscriptions of a specific actor_id
 
-        Access the indvidual subscriptions in .dbsubscriptions and the subscription data
-        in .subscriptions as a dictionary
+    Access the indvidual subscriptions in .dbsubscriptions and the subscription data
+    in .subscriptions as a dictionary
     """
 
     def fetch(self):
         if self.subscriptions is not None:
             return self.subscriptions
-        if not self.list:
+        if not self.list and self.config:
             self.list = self.config.DbSubscription.DbSubscriptionList()
-        if not self.subscriptions:
+        if not self.subscriptions and self.list:
             self.subscriptions = self.list.fetch(actor_id=self.actor_id)
         return self.subscriptions
 
@@ -144,23 +180,29 @@ class Subscriptions(object):
         if not self.list:
             logging.debug("Already deleted list in subscriptions")
             return False
-        for sub in self.subscriptions:
-            diff_list = self.config.DbSubscriptionDiff.DbSubscriptionDiffList()
-            diff_list.fetch(actor_id=self.actor_id, subid=sub["subscriptionid"])
-            diff_list.delete()
+        if self.subscriptions:
+            for sub in self.subscriptions:
+                if not self.config:
+                    continue
+                diff_list = self.config.DbSubscriptionDiff.DbSubscriptionDiffList()
+                diff_list.fetch(actor_id=self.actor_id, subid=sub["subscriptionid"])
+                diff_list.delete()
         self.list.delete()
         self.list = None
         self.subscriptions = None
         return True
 
-    def __init__(self,  actor_id=None, config=None):
-        """ Properties must always be initialised with an actor_id """
+    def __init__(self, actor_id=None, config=None):
+        """Properties must always be initialised with an actor_id"""
         self.config = config
         if not actor_id:
             self.list = None
             logging.debug("No actor_id in initialisation of subscriptions")
             return
-        self.list = self.config.DbSubscription.DbSubscriptionList()
+        if self.config:
+            self.list = self.config.DbSubscription.DbSubscriptionList()
+        else:
+            self.list = None
         self.actor_id = actor_id
         self.subscriptions = None
         self.fetch()
